feat: add create account
While doing this, I learnt about n-tier architecture and domain-driven
architecture. I also discovered the unique laws that guide data access
in Java 😅. Anyways, the learning experience is fun so far and I want
to see how farI go with this.

For those who may read this in the future, after a long series of intro-
spection, I have decided to take a step further in learning Java as part
of my plans to excel and reach higher echelons of life. From learning
Java to learning the Spring framework and SpringBoot, I want to be good
at them so that I can build big things with them and get well-paid jobs.
Doing this will also increase my understanding, knowledge and exposure
of things as a software engineer.

I'll write a READNE about my intentions of building this project in the
next commit.

A special commit to Victor Chigbo | Chivicks Hazard from the future:
In case you see all this, you're welcome 🤗❤💯
After everything we have been through, we won! I am so proud of you for
doing what you did to get to where you are right now.

I wish you all the best in the future to come.

With love,
Victor Chigbo | Chivicks Hazard.
Saturday, 20th September, 2025.