feat: implement airtime and data recharge
This was an easy one, just that the volume of code that I need to write
is much (but AI was very helpful with that). I noticed a UX flaw though:
Each time I still want to use a service, after selecting a service, I'll
put my pin back again. As someone who is primarily a frontend developer,
fixing this might take a lil while because of how I am going to
implement it. Well, at least I'm making progress 😁

Then, hardcoding the data plans is actually a bad practice. I just want
this to work which is why I did it like that. I think that the next
thing I will do it to put these plans in a database and fetch them when
necessary.

One thing on my mind that I want to learn with this project is caching,
like how to store and retrieve data without going back to the database
every time whenever we need it. This will help to make the app perform
faster and more efficiently.

So yeah. That's it for now.

See you in the next commit 👋